Break ConnectedSource into multiple source files
Fixes STSMACOM-97.
